Move router from App.js to separate file
