+++ frontend/smart-kickers-game/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './App';
+import Router from './router/Router';
 import { initLibs } from './appConfig';
 import GameDataContextProvider from './contexts/GameDataContext.js';
 
@@ -10,7 +10,7 @@ const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
     <GameDataContextProvider>
-      <App />
+      <Router />
     </GameDataContextProvider>
   </React.StrictMode>
 );
